Dissolver v0.9.3 Dissolver

Pagination for Ecto and Phoenix.

Dissolver is the continuation and fork of the fine work kerosene Out of respect to the authors of kerosene I won't be publishing this on https://hex.pm/ untill and if this becomes the replacement for kerosene.

Until then you will need to install this from github.

Installation

add Dissolver to your mix.exs dependencies:

def deps do
  [
    {:dissolver, github: 'MorphicPro/dissolver'}
  ]
end

Next provide Dissolver your Repo module via the config. Add the following to your config:

...
config :dissolver,
  repo: MyApp.Repo
  per_page: 2

import_config "#{Mix.env()}.exs"

For more information about the configuration options look at the Configurations section

Now you are ready to start using Dissolver.

Usage

Start paginating your queries

def index(conn, params) do
  {products, paginator} =
  Product
  |> Product.with_lowest_price
  |> Dissolver.paginate(params)

  render(conn, "index.html", products: products, paginator: paginator)
end

Add the view helper to your view

defmodule MyApp.ProductView do
  use MyApp.Web, :view
  import Dissolver.HTML
end

Generate the links using the view helper in your template

<%= paginate @conn, @paginator %>

Importing Dissolver.HTML provides your template access to Dissolver.HTML.paginate/3 as the prior example shows. The Dissolver.HTML.paginate/3 can take a host of options to aid the theme from how many links to show (window: integer) to what the link labels should read.

By default the theme used to generate the html is the Dissolver.HTML.Simple theme. It will only provide the very basic prev|next buttons. For more theme options, including providing your own, read the following Configurations

Configuration

This module uses the following that can be set as globals in your config/config.ex configurations

:repo - Required Your app's Ecto Repo

:theme (default: Dissolver.HTML.Simple) - A module that implements the Dissolver.HTML.Theme behavior There are a few pre defiend theme modules found in dessolver/html/

Dissolver.HTML.Simple - This is the default with only previous | next links

Dissolver.HTML.Bootstrap - A Bootstrap 4 theme

Dissolver.HTML.Foundation - A Foundation theme

Dissolver.HTML.Materialize - A Materialize theme

Dissolver.HTML.Semantic - A Semantic UI theme

Dissolver.HTML.Tailwind - A Tailwind CSS theme

:per_page (default: 10) - The global per page setting

:max_page - The limit of pages allow to navigate regardless of total pages found This option is ignored if not provided and defaults to total pages found in the query.

:lazy (default: false) - This option if enabled will result in all Dissolver.paginate/3 calls return an Ecto.Query rather than call Repo.all. This is useful for when you need to paginate on an association via a preload. TODO: provide example. ##

Functions

paginate(query, params, opts \\ [])

paginate(Ecto.Query.t(), map(), nil | keyword()) ::
  {list(), Dissolver.Paginator.t()}